langgraph/marketing_blog_agent/pyproject.toml
[project]
name = "marketing-blog-agent"
version = "0.0.1"
description = "Marketing blog writer agent using LangGraph with FAISS search and editing tools."
authors = [{ name = "Course Demo", email = "demo@example.com" }]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
dependencies = [
    # Core agent framework
    "langgraph>=0.2.6",
    "langgraph-cli[inmem]>=0.4.2",
    "langchain-core>=0.3.0",
    "langchain-anthropic>=0.2.0",
    # Utilities
    "python-dotenv>=1.0.1",
    "pydantic>=2.0.0",
    "numpy>=1.26",
    "faiss-cpu>=1.8.0.post1",
    # Embeddings via OpenAI (for FAISS index)
    "openai>=1.51.2",
]

[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["agent"]
[tool.setuptools.package-dir]
"agent" = "src/agent"

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    # We actually do want to import from typing_extensions
    "UP035",
    # Relax the convention by _not_ requiring documentation for every function parameter.
    "D417",
    "E501",
]
extend-exclude = [".venv", "**/.venv/**"]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
[tool.ruff.lint.pydocstyle]
convention = "google"

[dependency-groups]
dev = ["anyio>=4.7.0", "mypy>=1.13.0", "pytest>=8.3.5", "ruff>=0.8.2"]
